Upgrade MiniApp to React Native 0.49.0
