+++ package.json
@@ -6,8 +6,8 @@
     "start": "node node_modules/react-native/local-cli/cli.js start"
   },
   "dependencies": {
-    "react": "16.0.0-alpha.12",
-    "react-native": "0.48.3"
+    "react": "16.0.0-beta.5",
+    "react-native": "0.49.0"
   },
   "ern": {
     "version": "1000.0.0",
